requirements.txt
streamlit
tensorflow
numpy
pillow
scikit-learn
matplotlib
